fix: Refactor and clean up code across multiple files
- Updated logging function in Global.h to use localtime_s for safer time handling.
- Initialized pointers to nullptr in EndScreenState.h, IntroState.h, MainGameState.h, MapEditorState.h, MapSelectState.h, OptionsState.h, and TitleState.h to prevent potential dereferencing of uninitialized pointers.
- Added destructor to Tower class to manage memory for critterTargettingStrategy.
- Changed return types of getBuyingCost and getTotalTowersPlaced to const in Tower and TowerGroup classes respectively.
- Updated ParallaxBackground and ParallaxLayer to use float instead of int for dimensions and positions, improving precision.
- Removed unnecessary dynamic memory allocations and replaced them with stack allocations where applicable.
- Cleaned up memory management in various classes, ensuring proper deletion of dynamically allocated objects.
- Adjusted volume slider constructor to accept float parameters for position.

+++ src/critter/FastCritter.cpp
@@ -6,10 +6,10 @@ FastCritter::FastCritter(int level, SDL_FRect start, Map* map)
 	speed += level * SPEED_SCALE_AMOUNT;
 	maxHitPoints += level * HITPOINT_SCALE_AMOUNT;
 	hitPoints = maxHitPoints;
-	reward += level * REWARD_SCALE_AMOUNT;
+	reward += static_cast<int>(level * REWARD_SCALE_AMOUNT);
 
 	setupAnimationFrames();
-	frameTime = (1.0f / animationFramesWalkSide.size()) * (0.5 * (1 + Critter::SPEED_SCALE_FACTOR));
+	frameTime = (1.0f / animationFramesWalkSide.size()) * (0.5f * (1 + Critter::SPEED_SCALE_FACTOR));
 	//std::cout << std::fixed << "Fast Critter Frame Time: " << frameTime << std::endl;
 
 }
